backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
google-cloud-aiplatform==1.38.1
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
google-generativeai==0.3.2
python-dotenv==1.0.0
asyncio==3.4.3
aiofiles==23.2.1
httpx==0.25.2
